The RPG sample now builds, and _begins_ running, but fails when you start to run the game.

+++ MonoGameSamples/RolePlayingGame/RolePlayingGame/MenuScreens/LoadingScreen.cs
@@ -179,7 +179,7 @@ namespace RolePlaying
                 Viewport viewport = ScreenManager.GraphicsDevice.Viewport;
                 Vector2 viewportSize = new Vector2(viewport.Width, viewport.Height);
 
-                Color color = new Color(255, 255, 255, TransitionAlpha);
+                Color color = new Color(255, 255, 255, (int)TransitionAlpha);
 
                 spriteBatch.Begin();
                 spriteBatch.Draw(loadingBlackTexture, loadingBlackTextureDestination,
